refactor; Moved out fixtures for reuse and approval client into its own module
